Make unprivileged cancelation of priviledged process work
Mostly removing a code -- we really ought to keep this simple.

Trying to send the TERMinate signal was very polite of us.
However, we need to also handle the case, when some thirdparty
signals to us in a less polite manner.

Hence, oscap-workbench-wrapper.sh shall finish when any of the
parents is drawn down regardless of the cause. The wrapper shall
not check the input but the availability of input (stdin). For
the wrapper there is no meaning of life without stdin.

That leaves us with oscap-workbench.sh which purpose has been
reduced and it could be rewritten to C++. In future release
we could also make oscap-workbench.sh close the stdin of
oscap-workbench-wrapper.sh when a TERM signal is received.