Fixes the long time to open TailoringWindow in MacOS
There seems to be a bug in Qt (MacOS implementation) when it tries to calculate geometry/draw tri-state checkboxes that makes it spend much more time than usual.

This commit changes the traversal of the tree from preorder to postorder so that we set the expanded items while they are still invisible, thus avoiding the geometry calculation and drawing. Only when the top-most item is expanded that all items are drawed, once.

Signed-off-by: Wesley Ceraso Prudencio <wcerasop@redhat.com>