Generate result-based remediation from tailored profile
Users can generate remediation script from scan results
of a tailored profile.

Unfortunately, the current design of SCAP Workbench doesn't allow
a clear way of doing this. The scan is run in a separated oscap
process. SCAP Workbench doesn't have access to oscap internal
xccdf_session which creates the ARF. It can't obtain the Tailoring
component reference ID.

Instead, we will save the tailoring document to a temporary file
and use the temporary file when generating the remediation.

Resolves: RHBZ#1640715