Allow profile description customization
When opening the profile editor, the description field is populated but
is marked ReadOnly for some reason; all the back-end code existed to
make this read-write (except in the case when the profile lacks a
description, see the comment in TailoringWindow::setProfileDescription
and TailoringWindow::setProfileTitle).

This makes the field read-write.

Note that the behavior is less than ideal: <xccdf:description /> can
include HTML content, and while this is a HTML-aware editor, hitting the
return key inserts merely a '\n' into the text stream and not a <br />
(or better, <xhtml:br /> with XML-NS awareness) element.

Perhaps this behavior should be corrected at the openscap library level,
detecting the HTML XML namespace applied on the <xccdf:description />
element and namespacing the input correctly?

Signed-off-by: Alexander Scheel <alex.scheel@canonical.com>