Fix remote scanning with sudo
This fix ensures that various temp files that are created and used in the process
have the same owner all the time - if the scan is executed with sudo,
then the owner of report, ARF and result files has to be root from the
beginning to the end.

The root user in RHEL9 can't just write to files owned by other users.